Fix new data transfer infinite looping with transfer size >2GB
get_new_data calculates size with a long, then passing this to receive_new_data which uses an int and has a while >0 loop. Sizes >2GB result in a negative value of the int used in receive_new_data, causing an infinite loop.

+++ blockimg/blockimg.cpp
@@ -345,7 +345,7 @@ struct NewThreadInfo {
     pthread_cond_t cv;
 };
 
-static bool receive_new_data(const unsigned char* data, int size, void* cookie) {
+static bool receive_new_data(const unsigned char* data, long size, void* cookie) {
     NewThreadInfo* nti = reinterpret_cast<NewThreadInfo*>(cookie);
 
     while (size > 0) {
